# INLINABLE insert and delete, bench, small doc additions.