--- a/bench/Cuckoo.hs
+++ b/bench/Cuckoo.hs
@@ -39,7 +39,11 @@ main = do
             [ insertBench 10000 12000 [1..10000]
             ]
         , bgroup "member"
-            [ memberBench 10000 cf10k [5000..15000]
+            [ memberBench "1:1 hit:hiss" 10000 cf10k (interleave [5000..10000] [10001..15000])
+            , memberBench "only hit" 10000 cf10k [1..10000]
+            ]
+        , bgroup "delete"
+            [ deleteBench "all" 10000 cf10k [1..10000]
             ]
         ]
 
@@ -55,7 +59,22 @@ insertBench n cap xs0 = bench ("n-" <> show n <> "-cap-" <> show cap) $ whnf f x
         failed <- filterM (fmap not . insert cf) xs
         pure $! length failed
 
-memberBench :: Int -> CuckooFilter (PrimState IO) 4 10 Int -> [Int] -> Benchmark
-memberBench n cf xs = bench (show n) $ whnfIO f
+memberBench :: String -> Int -> CuckooFilter (PrimState IO) 4 10 Int -> [Int] -> Benchmark
+memberBench label n cf xs = bench (label <> " - " <> show n) $ whnfIO f
   where
     f = fmap length (filterM (member cf) xs)
+
+deleteBench :: String -> Int -> CuckooFilter (PrimState IO) 4 10 Int -> [Int] -> Benchmark
+deleteBench label n cf0 xs = bench (label <> " - " <> show n) $ whnfIO f
+  where
+    f = do
+        (cf, _) <- splitCuckooFilter cf0
+        fmap length (filterM (delete cf) xs)
+
+
+
+-- | Doesn't cause actual difference in bench.
+interleave :: [a] -> [a] -> [a]
+interleave (a:as) (b:bs) = a:b:interleave as bs
+interleave as bs = as <> bs
+