--- a/cuckoo.cabal
+++ b/cuckoo.cabal
@@ -166,3 +166,22 @@ benchmark internal-benchmarks
         , base >=4.10 && <5
         , criterion >= 1.5
 
+benchmark cuckoo-benchmarks
+    type: exitcode-stdio-1.0
+    hs-source-dirs: bench
+    main-is: Cuckoo.hs
+    default-language: Haskell2010
+    ghc-options:
+        -Wall
+        -rtsopts
+        -threaded
+        -with-rtsopts=-N
+    build-depends:
+        -- internal
+        , cuckoo
+
+        -- external
+        , base >=4.10 && <5
+        , criterion >= 1.5
+        , primitive
+