Added benchmarks for the main public interface ops, also marked insert and delete INLINABLE, so the callsite can optimize further. This had some small but noticable impact on the benchmark.

```
insert: 16ms -> 11ms
delete: 26ms -> 16ms

```